requirements.txt
##### Requirements for OPEN Controller #####

Pillow~=7.2.0
pyserial~=3.4